+++ x-pack/plugins/security_solution/server/lib/detection_engine/signals/utils.ts
@@ -26,8 +26,7 @@ import { BuildRuleMessage } from './rule_messages';
 import { parseScheduleDates } from '../../../../common/detection_engine/parse_schedule_dates';
 import { hasLargeValueList } from '../../../../common/detection_engine/utils';
 import { MAX_EXCEPTION_LIST_SIZE } from '../../../../../lists/common/constants';
-import { ShardError } from '../../types';
-
+import { ShardError } from '../../../../common/detection_engine/types';
 interface SortExceptionsReturn {
   exceptionsWithValueLists: ExceptionListItemSchema[];
   exceptionsWithoutValueLists: ExceptionListItemSchema[];
