+++ x-pack/plugins/security_solution/server/lib/hosts/types.ts
@@ -17,7 +17,8 @@ import {
   DocValueFieldsInput,
 } from '../../graphql/types';
 import { FrameworkRequest, RequestOptionsPaginated } from '../framework';
-import { Hit, Hits, SearchHit, TotalValue } from '../types';
+import { Hit, Hits } from '../types';
+import { SearchHit, TotalValue } from '../../../common/detection_engine/types';
 
 export interface HostsAdapter {
   getHosts(req: FrameworkRequest, options: HostsRequestOptions): Promise<HostsData>;
