+++ x-pack/plugins/security_solution/server/lib/types.ts
@@ -16,7 +16,6 @@ import { Sources } from './sources';
 import { Note } from './note/saved_object';
 import { PinnedEvent } from './pinned_event/saved_object';
 import { Timeline } from './timeline/saved_object';
-import { SearchTypes } from './detection_engine/signals/types';
 
 export * from './hosts';
 
@@ -39,94 +38,6 @@ export interface SiemContext {
   context: RequestHandlerContext;
   user: AuthenticatedUser | null;
 }
-
-export interface TotalValue {
-  value: number;
-  relation: string;
-}
-
-export interface BaseHit<T> {
-  _index: string;
-  _id: string;
-  _source: T;
-}
-
-export interface SearchResponse<T> {
-  took: number;
-  timed_out: boolean;
-  _scroll_id?: string;
-  _shards: ShardsResponse;
-  hits: {
-    total: TotalValue | number;
-    max_score: number;
-    hits: Array<
-      BaseHit<T> & {
-        _type: string;
-        _score: number;
-        _version?: number;
-        _explanation?: Explanation;
-        fields?: string[];
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        highlight?: any;
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        inner_hits?: any;
-        matched_queries?: string[];
-        sort?: string[];
-      }
-    >;
-  };
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  aggregations?: any;
-}
-
-export interface EqlSequence<T> {
-  join_keys: SearchTypes[];
-  events: Array<BaseHit<T>>;
-}
-
-export interface EqlSearchResponse<T> {
-  is_partial: boolean;
-  is_running: boolean;
-  took: number;
-  timed_out: boolean;
-  hits: {
-    total: TotalValue;
-    sequences?: Array<EqlSequence<T>>;
-    events?: Array<BaseHit<T>>;
-  };
-}
-
-export interface ShardsResponse {
-  total: number;
-  successful: number;
-  failed: number;
-  skipped: number;
-  failures?: ShardError[];
-}
-
-export interface ShardError {
-  shard: number;
-  index: string;
-  node: string;
-  reason: {
-    type: string;
-    reason: string;
-    index_uuid: string;
-    index: string;
-    caused_by: {
-      type: string;
-      reason: string;
-    };
-  };
-}
-
-export interface Explanation {
-  value: number;
-  description: string;
-  details: Explanation[];
-}
-
-export type SearchHit = SearchResponse<object>['hits']['hits'][0];
 
 export interface TermAggregation {
   [agg: string]: {
